src/InjectorTest.java
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
class InjectorTest {
    @Test
    void testInject() throws IllegalAccessException {
        // Create an object to be injected
        SomeClass obj = new SomeClass();

        // Create a mock dependency object
        DependencyClass dependency = Mockito.mock(DependencyClass.class);

        // Create an instance of the Injector
        Injector injector = new Injector();

        // Inject the mock dependency into the object
        SomeClass result = injector.inject(obj, dependency);

        // Perform assertions to verify the injection
        Assertions.assertNotNull(result.getDependency());
        Assertions.assertEquals(dependency, result.getDependency());
    }

    // Classes for testing
    static class SomeClass {
        @AutoInjectable
        private DependencyClass dependency;

        public DependencyClass getDependency() {
            return dependency;
        }
    }

    static class DependencyClass {
        // Dependency implementation
    }
}